Reduce the output of the test macros
